refactor - added event constants and tidied naming
