+++ src/app.js
@@ -2,7 +2,7 @@ import socketio from './socketIO/socketio';
 const express = require('express');
 const http = require('http');
 
-const port = 3000;
+const port = 3001;
 
 // create express/http server
 const app = express();
