+++ src/socketio.js/eventConstants.js
@@ -1,0 +1,4 @@
+const PREGAMESTART = "start-pre-game";
+const PREGAMEFINISH = "finish-pre-game";
+const CARDS = "cards";
+const CONTINUE = "continue";
